requirements.txt
djangorestframework
django-filter
django
django-rest-framework